% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.car}
\alias{greta.car}
\title{Conditional Autoregressive Models in greta}
\description{
Distributions and utility functions that extend greta by making
it easy to define conditional autoregressive models.
}
\examples{

# add a simple example here to introduce the package!

}
